requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

langchain==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0

pinecone==5.0.0
python-dotenv==1.0.1
pydantic==2.8.2
